feat(players): add sorting controls
